requirements.txt
# Core chess libraries
python-chess==1.999
stockfish==3.28.0
tqdm==4.66.1